hands_for_brain/requirement.txt
google-generativeai
playwright
python-dotenv
anthropic